fix fixture lifetimes
