+++ test/NLightning.Bolts.Tests/Docker/Fixtures/LightningRegtestNetworkFixture.cs
@@ -92,3 +92,11 @@ public class LightningRegtestNetworkFixture : IDisposable
         }
     }
 }
+
+[CollectionDefinition("regtest")]
+public class LightningRegtestNetworkFixtureCollection : ICollectionFixture<LightningRegtestNetworkFixture>
+{
+    // This class has no code, and is never created. Its purpose is simply
+    // to be the place to apply [CollectionDefinition] and all the
+    // ICollectionFixture<> interfaces.
+}
