+++ test/NLightning.Bolts.Tests/Docker/PostgresTests.cs
@@ -1,3 +1,4 @@
+
 namespace NLightning.Bolts.Tests.Docker;
 
 using Fixtures;
@@ -5,7 +6,8 @@ using Utils;
 using Xunit.Abstractions;
 
 #pragma warning disable xUnit1033 // Test classes decorated with 'Xunit.IClassFixture<TFixture>' or 'Xunit.ICollectionFixture<TFixture>' should add a constructor argument of type TFixture
-public class PostgresTests : IClassFixture<PostgresFixture>
+[Collection("postgres")]
+public class PostgresTests 
 {
     private readonly PostgresFixture _postgresFixture;
